requirements.txt
pyyaml>=6.0
cryptography>=41.0
requests>=2.31.0
jinja2>=3.1.2
# Packaging is useful for version comparison in delta logic if needed
packaging>=23.0
